型、変数名を若干変更

+++ game/game/source_code/application/object/objects/sprite/damage_effect.h
@@ -29,7 +29,7 @@ public:
     virtual ~DamageEffect();
     void Update();
     void Draw();
-	void SetHP(const int hp){
+	void SetHP(const float hp){
 		hp_ = hp;
 	}
 private:
@@ -45,7 +45,9 @@ private:
     FLASH_MODE flash_mode_;
     float flash_alpha_;
 
-    int hp_;
+	//-------------------------------------
+	// ‘Ì—Í
+    float hp_;
 };
